Tweak orm for tool call column and fix some actor bugs
