+++ alembic/versions/9a505cc7eca9_create_a_baseline_migrations.py
@@ -97,7 +97,7 @@ def upgrade() -> None:
         sa.Column("text", sa.String(), nullable=True),
         sa.Column("model", sa.String(), nullable=True),
         sa.Column("name", sa.String(), nullable=True),
-        sa.Column("tool_calls", letta.metadata.ToolCallColumn(), nullable=True),
+        sa.Column("tool_calls", letta.orm.message.ToolCallColumn(), nullable=True),
         sa.Column("tool_call_id", sa.String(), nullable=True),
         sa.Column("created_at", sa.DateTime(timezone=True), nullable=True),
         sa.PrimaryKeyConstraint("id"),
